@@ -1,8 +1,6 @@
 #Conferences are the main administrative resource, they define a bunch of settings
 #and own conference items (which are what users)
 class Conference < ActiveRecord::Base
-  attr_accessible :conference_year, :tax_rate, :conference_items_attributes, :active
-
   #associations
   has_many :conference_items
   has_many :line_items, :through => :conference_items
@@ -26,28 +24,6 @@ class Conference < ActiveRecord::Base
   after_save :enforce_one_active_conference
   after_create :build_schedule
 
-
-  #returns the number of proposals scoped to this conference
-  def proposal_count
-    proposals.count
-  end
-
-  def accepted_proposal_count
-    proposals.accepted.count
-  end
-
-  def wait_listed_proposal_count
-    proposals.wait_listed.count
-  end
-
-  def registration_count
-    line_items.where(:paid => true).count
-  end
-
-  def pending_registration_count
-    line_items.where(:paid => false).count
-  end
-
   def registration_breakdown
     conference_items.map do |item|
       {
@@ -57,13 +33,13 @@ class Conference < ActiveRecord::Base
     end
   end
 
-  # essentially a named scope, but implemented as a class method beecause scopes must be chainable and find_by_x isn't chainable
+  # essentially a named scope, but implemented as a class method beecause scopes must be chainable and find_by isn't chainable
   def self.active
-    find_by_active(true)
+    find_by(active: true)
   end
 
   def to_param
-    conference_year
+    conference_year.to_s
   end
 
   private
