@@ -1,10 +1,6 @@
-# Read about factories at https://github.com/thoughtbot/factory_girl
-
 FactoryGirl.define do
-  factory :country do
-    factory :usa do
-        name "USA"
-        category 1
-    end
+  factory :country, aliases: [:usa] do
+    name 'USA'
+    category 1
   end
 end
