@@ -4,14 +4,14 @@ describe Discount do
 
   describe :validations do
     it 'should not allow two discounts with the same key' do
-      discount = FactoryGirl.create(:discount)
-      discount2 = FactoryGirl.build(:discount)
+      discount = create(:discount)
+      discount2 = build(:discount)
       discount2.discount_key = discount.discount_key
       expect(discount2).to be_invalid
     end
 
     it 'should have a discount key of length six' do
-      discount = FactoryGirl.build(:discount)
+      discount = build(:discount)
       discount.discount_key = '123'
       expect(discount).to be_invalid
     end
@@ -31,7 +31,7 @@ describe Discount do
     end
 
     it 'should build a price for each conference item' do
-      conference = FactoryGirl.create(:conference_with_3_items)
+      conference = create(:conference_with_items)
       discount = conference.discounts.build
 
       discount.build_prices_for_conference_items
