@@ -8,19 +8,19 @@ describe Presenter do
     end
 
     it 'should return Registered when itinerary has a registered conference line item' do
-      itinerary = FactoryGirl.create(:itinerary_with_paid_conference)
+      itinerary = create(:itinerary_with_item, paid: true)
       status = Presenter.presenter_conference_status(itinerary.user.email, itinerary.conference.id)
       expect(status).to eql('Registered')
     end
 
     it 'should return Pending Registration when itinerary has a pending registered line item' do
-      itinerary = FactoryGirl.create(:itinerary_with_pending_conference)
+      itinerary = create(:itinerary_with_item, paid: false)
       status = Presenter.presenter_conference_status(itinerary.user.email, itinerary.conference.id)
       expect(status).to eql('Pending Registration')
     end
 
     it 'should return Not Registered when itinerary does not have line items' do
-      itinerary = FactoryGirl.create(:itinerary)
+      itinerary = create(:itinerary)
       status = Presenter.presenter_conference_status(itinerary.user.email, itinerary.conference.id)
       expect(status).to eql('Not Registered')
     end
