@@ -25,8 +25,8 @@ describe 'SignUp' do
     expect(response).to render_template('users/edit')
 
     user = assigns(:user)
-    country = FactoryGirl.create(:country)
-    put "users/#{user.id}", { user: {
+    country = create(:country)
+    put "/users/#{user.id}", { user: {
       first_name: 'Jane',
       last_name: 'Doe',
       nametag_name: 'Jane Doe',
