@@ -7,7 +7,6 @@ Spork.prefork do
   ENV["RAILS_ENV"] ||= 'test'
   require File.expand_path("../../config/environment", __FILE__)
   require 'rspec/rails'
-  require 'rspec/autorun'
   # Requires supporting ruby files with custom matchers and macros, etc,
   # in spec/support/ and its subdirectories.
   Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
@@ -49,4 +48,6 @@ RSpec.configure do |config|
   config.include ValidUserRequestHelper, :type => :request
 
   config.infer_spec_type_from_file_location!
+  config.include Requests::JsonHelpers, type: :request
+  config.include FactoryGirl::Syntax::Methods
 end
