@@ -3,16 +3,16 @@ module ValidUserRequestHelper
 
   # for use in request specs
   def sign_in_as_a_valid_user
-    @user ||= FactoryGirl.create :user
+    @user ||= create(:user)
     post_via_redirect user_session_path, 'user[email]' => @user.email, 'user[password]' => @user.password
   end
 
   def sign_in_as_a_admin_user
-    @user ||= FactoryGirl.create :admin
+    @user ||= create(:user, :admin)
     post_via_redirect user_session_path, 'user[email]' => @user.email, 'user[password]' => @user.password
   end
 
   def create_conference
-    FactoryGirl.create :conference
+    create(:conference)
   end
 end
